+++ AddOns/DBM-Raids-Vanilla/BlackrockDepths/GeneralAngerforge.lua
@@ -1,10 +1,10 @@
 if DBM:GetTOC() < 110005 then return end
-local mod	= DBM:NewMod(2668, "DBM-Raids-WarWithin", 2, 1301)
+local mod	= DBM:NewMod(2668, "DBM-Raids-Vanilla", 5, 1301)
 local L		= mod:GetLocalizedStrings()
 
 mod.statTypes = "lfr,normal,heroic"
 
-mod:SetRevision("20240909033944")
+mod:SetRevision("20240916054251")
 mod:SetCreatureID(226316)
 mod:SetEncounterID(3045)
 --mod:SetUsedIcons(8, 7, 6)