+++ AddOns/DBM-Raids-Vanilla/BlackrockDepths/TheSeven.lua
@@ -1,10 +1,10 @@
 if DBM:GetTOC() < 110005 then return end
-local mod	= DBM:NewMod(2667, "DBM-Raids-WarWithin", 2, 1301)
+local mod	= DBM:NewMod(2667, "DBM-Raids-Vanilla", 5, 1301)
 local L		= mod:GetLocalizedStrings()
 
 mod.statTypes = "lfr,normal,heroic"
 
-mod:SetRevision("20240908020318")
+mod:SetRevision("20240916054251")
 mod:SetCreatureID(226307, 226310, 226309, 226313, 226311, 226312, 226308)--226307 Anger'rel, 226310/doomrel, 226309/doperel, 226313/gloomrel, 226311/haterel, 226312/seethrel, 226308/vilerel
 mod:SetEncounterID(3048)
 mod:SetUsedIcons(1, 2)