+++ AddOns/TomCats/hallowsend/MainRoute.lua
@@ -243,7 +243,7 @@ end
 --    TomCats:Register(
 --        {
 --            name = "@shortName@",
---            version = "2.5.43",
+--            version = "2.5.44",
 --        }
 --    )
 --end