+++ AddOns/DBM-Core/CHANGELOG.md
@@ -1,10 +1,24 @@
 # DBM - Core
 
-## [11.0.10](https://github.com/DeadlyBossMods/DeadlyBossMods/tree/11.0.10) (2024-09-13)
-[Full Changelog](https://github.com/DeadlyBossMods/DeadlyBossMods/compare/11.0.9...11.0.10) [Previous Releases](https://github.com/DeadlyBossMods/DeadlyBossMods/releases)
+## [11.0.13](https://github.com/DeadlyBossMods/DeadlyBossMods/tree/11.0.13) (2024-09-17)
+[Full Changelog](https://github.com/DeadlyBossMods/DeadlyBossMods/compare/11.0.11...11.0.13) [Previous Releases](https://github.com/DeadlyBossMods/DeadlyBossMods/releases)
 
-- prep new tag  
-- extend P3 data on heroic queen  
-    Fix invalid spellid/name for null detonation timer and alerts  
-- Change venomous rain to just warn EVERYONE to spread on ocast, since we can't detect who precircles are on until it's too late.  
+- disable that last minute change for now, it has load order issues  
+- bump alpha  
+- Add option to disable built in music options. Closes https://github.com/DeadlyBossMods/DBM-SoundEventsPack/issues/5  
+- Update localization.ru.lua (#1244)  
+- Update koKR (#1243)  
+- Fixed a few places short names weren't being applied.  
+- Queen Ansurek Update:  
+     - Improved Wrest a little bit  
+     - Changed Reactive Toxin target message to give count and no longer be filtered  
+     - Added icon setting and yell ptions to reactive toxin that support both matching BW or doing it ascended method or disabling globally  
+     - Changed icons for Royal Condemnation to be more compatible  
+     - Added queens summon icon marking using remaining icons  
+     - Fixed bug where Gloom Eruption announce/timer was started in wrong spot.  
+    Blood Twister Ovinax Update:  
+     - Fixed variable names that caused raid leader overide to not be set correctly.  
+    Both of above:  
+     - Icon dropdown override menus are now categorized as "icon" options and not "misc" so they are grouped properly if user is using the "show icons as separate category" option.  
+- messaging tweaks  
 - bump alpha